+++ src/main/java/com.example.demo/domain/dto/ProductRowDto.java
@@ -7,6 +7,6 @@ public class ProductRowDto {
 
     // Cena i nazwa poszczególnego produktu
     private String name;
-    private Double netto;
-    private Double brutto;
+    private Float netto;
+    private Float brutto;
 }
